requirements.txt
Django==5.1.2
django-cors-headers==4.5.0
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
